requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
jinja2
python-multipart
python-docx
httpx